+++ ports/raspberrypi/boards/raspberry_pi_pico_w/mpconfigboard.mk
@@ -12,7 +12,6 @@ CIRCUITPY__EVE = 1
 
 CIRCUITPY_CYW43 = 1
 CIRCUITPY_SSL = 1
-CIRCUITPY_SSL_MBEDTLS = 1
 CIRCUITPY_HASHLIB = 1
 CIRCUITPY_WEB_WORKFLOW = 1
 CIRCUITPY_MDNS = 1
