+++ shared-bindings/ssl/SSLSocket.h
@@ -27,7 +27,11 @@
 #ifndef MICROPY_INCLUDED_SHARED_BINDINGS_SSL_SSLSOCKET_H
 #define MICROPY_INCLUDED_SHARED_BINDINGS_SSL_SSLSOCKET_H
 
+#if CIRCUITPY_SSL_MBEDTLS
+#include "shared-module/ssl/SSLSocket.h"
+#else
 #include "common-hal/ssl/SSLSocket.h"
+#endif
 
 extern const mp_obj_type_t ssl_sslsocket_type;
