+++ shared-bindings/ssl/__init__.h
@@ -27,7 +27,11 @@
 #ifndef MICROPY_INCLUDED_SHARED_BINDINGS_SSL___INIT___H
 #define MICROPY_INCLUDED_SHARED_BINDINGS_SSL___INIT___H
 
+#if CIRCUITPY_SSL_MBEDTLS
+#include "shared-module/ssl/SSLContext.h"
+#else
 #include "common-hal/ssl/SSLContext.h"
+#endif
 
 void common_hal_ssl_create_default_context(ssl_sslcontext_obj_t *self);
